requirements.txt
flask
gunicorn
pandas
prophet
scikit-learn
pypdf
requests
python-dotenv